refactor: freed the GolemNetwork name by changing the original implmentation to JobManager
