+++ src/experimental.ts
@@ -1,4 +1,3 @@
 export * from "./experimental/job";
-export * from "./experimental/golem_network";
 export * from "./experimental/reputation";
 export * from "./experimental/new-api";
