+++ src/experimental/job/index.ts
@@ -1,1 +1,2 @@
 export { Job, JobState } from "./job";
+export { JobManager, JobManagerConfig } from "./job_manager";
